requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==4.2
django-allauth==65.1.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-js-asset==3.0.1
django-rest-auth==0.9.5
django-rest-knox==5.0.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
openapi-codec==1.3.2
packaging==24.2
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
six==1.17.0
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
